feat(indexer): Postgres/CloudSpanner/ClickHouse ID compatibility (#37717)
This change ensures the generated IDs fit in a 63 bit integer, and the
leading bit of the get_id() function is always zero.

For CloudSpanner, the codec now reverses the bits, shifts to the left by
one and subtracts 2^63 in order to produce a well distributed signed
64 bit int.

For Postgres, we cannot do the same since data is already present in the
table and we cannot change the mapping. But this is not a problem since
our ID generator will never produce numbers larger than 2^63.