ref(perf): Add duplicate detection on group hash (#37787)
This uses the existing grouping strategy hash already on the spans for http spans, since we have more refined rules around auto-parameterizing the description etc. If this works, it's also a better fingerprint to use since it allows us to directly query span.groups etc. on a transaction (eg. suspect spans).

The one issue with this approach is that it only works after manager.save is called, which is where span groups gets calculated, so if we need to propagate data into the transaction based on detection we'll have to figure something else out (or move this to 'save')