feat(perf): Add detection for render-blocking asset performance issues
Tag transactions that have slow asset load spans before a slow FCP as having
render-blocking assets. The thresholds are configurable, but currently we're
looking for transactions with an FCP between 2s and 10s, where an asset load
takes up at least 25% of that time.

The thresholds will be tuned before we start generating
actual Performance Issues from this data - tagging the transactions will let us
see what we're detecting it and validate/tune it before it becomes visible to
users.

This detector's use of event properties is a little awkward given the current
`PerformanceDetector` interface, but I thought it would be better to get the
rest of our planned detectors in before we refactor too much.

Fixes PERF-1677