fix(tests): Fixing some backend test flakes
Successively calling before_now results in some flakes
particularly in tests that are calculating aggregates
on transaction.duration. Introduced a util that takes
a datetime object and timedelta kwargs and returns
a datetime object with the delta applied. It's kinda
like the before_now util except it's before any datetime
the user provides.

parameters