fix(mep): No compatible projects MEP alert
This modifies the conditions for showing 'only view compatible projects' on the 'incompatible with server side sampling alert'. Previously it was possible if you enabled dynamic sampling but none of your projects were compatible to see this message, but clicking on the link would not do anything since you have no compatible projects to switch to. This changes the alert to only show the 'update sdk' CTA since that is the only action the user can do to start seeing metric-side data.