fix(mep): Maintain fallback decision to txn summary
This maintains the fallback decision by re-using the existing provider on the transaction summary page. Still leaving the count check in since it's a simple way to know which side of data to show and gets around the fact that we're not awaiting the potential reloading of this provider, which can happen when going from a multi project view to a single project view.