ref(perf): Start replacing *DiscoverQuery with RQ
This is an example of replacing our fetch-as-component pattern with hooks, as was eventually intended. In this case using RQ to try it out on the performance part of the product, since almost every single request we make is through *DiscoverQuery so it would be easy to test it out with this hook and swap up RQ for another approach under the hood if we'd like to at a later point.