fix(replays): Read from `replayRecord` to show errors and duration in the details header (#38032)
Previously we were fetching the actual error objects from the `/eventsv2/` api, and then counting them up. 
This is still broken; it's probably the same root cause as [38002](https://github.com/getsentry/sentry/issues/38002), so it'll get fixed. Also we'll switch from `/eventsv2/` to the supported events api in #37998

So lots of things related, but this is a narrow fix and we get to remove some prop-drilling in the process.

Fixes #38001