fix(replays): Fix switching between orgs when starting on a replay page (#38029)
Turns out that `:orgId` in routes.tsx is a special name and must be used. There's a wrapper around all the `/organization/:orgId/*` routes
- Wrapper: https://github.com/getsentry/sentry/blob/master/static/app/routes.tsx#L1722
- Component: https://github.com/getsentry/sentry/blob/master/static/app/views/organizationDetails/index.tsx#L14-L24

`OrganizationDetails` tracks when the organization changes and makes sure the page-wide context/stores are reset or updated with the fresh values.

When we renamed from `:orgId` to `:ordSlug` that `<OrganizationDetails>` wasn't able to do its job anymore. Don't rename that param.

Fixes #37928