feat(perf_issues): Support multiple group ids in `Event` and `EventStream` (#38120)
This adds a basic `group_ids` property to `Event`. This is used in `EventStream` to pass the
`group_ids` associated with a performance issue to snuba.

We don't add these to the Kafka message headers in `EventStream` yet. These are used for passing
information about a group to the post process forwarder. The format of these will need to change
going forward to have some kind of mapping of group -> post_process_group info.