ref(metrics-indexer): Turn process_messages into MessageProcessor (#38225)
**context:**
addressing the comments in https://github.com/getsentry/sentry/pull/37745 specifically this comment https://github.com/getsentry/sentry/pull/37745#discussion_r949776817.

This change will allow me to initialize the indexer _not_ on every batch, but instead on when the strategy is created. If this would be recreating the indexer instance too often, we could move it farther up (aka when the factory is created vs the strategy). 

My idea is that in https://github.com/getsentry/sentry/pull/37745 we can add a `indexer_db_options` to the config which can be used to initialize the indexer. 

Something like:

```python
STORAGE_TO_INDEXER: Mapping[IndexerStorage, StringIndexer] = {
    IndexerStorage.CLOUDSPANNER: CloudSpannerIndexer,
    IndexerStorage.POSTGRES: PostgresIndexer,
    IndexerStorage.MOCK: MockIndexer,
}

class MessageProcessor:
    def __init__(self, config: MetricsIngestConfiguration):
        self._config = config
        self._indexer = STORAGE_TO_INDEXER[config.db_backend](**config.indexer_db_options)