refs(event): Partial typing for `Event` (#38128)
This adds partial typing to `sentry.eventstore.models`. Adding full typing was difficult here, since
`Group` is optional but a lot of places expect `event.group` to always exist.

`memoize` also needs to have typing added and I didn't want to tackle that yet. I also left a lot of
grouping config things untyped.

Once we move `group` out of event and into `GroupedEvent` it will be easier to finish up typing
here. Having the basic types in place should be helpful as we work on this anyway.