feat(replays): add download functionality to replay recording segment index (#38219)
Right now the frontend has to make an individual request to download every single recording segment on the details page. It's possible/common that there's over 100 segments for a recording, which means the frontend has to make _a lot_ of requests.

This PR adds functionality to the index endpoint so that it can be used to download a range of segments, using our standard pagination parameters / headers. The endpoint will also deflate the data and return it as JSON, which will remove the need for us to include `pako` client side. 

We can experiment with optimal loading of a replay details page by fetching only the first few segments and then the rest, or simply batching requests using a fixed offset (say grab 20 recording segments per request). 

We could also try including metadata about `checkout` segments, and fetch from checkout to checkout, but this will be more involved since that requires a client sdk change.