ref(hybrid): Standardize nomenclature around silos
Having previously wished to avoid committing to the term "silo" in
production code and instead used the terms "server component" and
"server mode", standardize naming of classes and variables on the terms
"silo" and "silo mode".

Also clean up some related names around silo mode decorators, and use
the terms "Model" and "Endpoint" (in reference to the decorated classes)
in preference to synonyms such as "data", "API", and "view".