fix(replays): Dont query for errors events when there are no ids in the condition list (#38266)
Previously we were always doing a query for errors, even if there were no errors to search for. This was returning lots of items because there was nothing to limit the response.

Now we do a search for `errorIds`, if there are items in the list. If the list is empty, do not search.

Fixes #38243

Depends on #38272