ref(writes-limiter): Add WriteLimiterFactory (#38192)
Addressing https://github.com/getsentry/sentry/pull/37745#discussion_r949779591 in https://github.com/getsentry/sentry/pull/37745

Right the `check_writes_limits` does both checking the limits, but also initializes a new `RedisSlidingWindowRateLimiter`. This causes some complexity because we end up looking up the config in order to get the options stored on the config in order for the redis rate limiter to be initialized. 

Instead I've pulled the logic out of initializing the rate limiters into the `WritesLimiterFactory`, and `get_ratelimiter` will get the right rate limiter based on the config. This means that each `WritesLimiter` has it's own redis rate limiter, and the factory is in charge of mapping the namespace to the `WritesLimiter` instances. 

This means as we add parameters like `db_backend` to be used to get the config, we don't have to pass those along to the rate limiter.