feat(hybrid): Scripts for silo mode decorators
The goal is for the scripts at `./scripts/silo` to be the primary source
of truth for which decorators belong on which classes. New decorators
should be added by running `./scripts/silo/all_decorators.sh`, not by
inserting them by hand.

Add test decorators in src.sentry.testutils that will cause tests to be
run or skipped in the appropriate silo mode. Currently, these tests
should not modify behavior when in monolith mode.