chore: Break up widgetBuilder.spec.tsx into multiple test files  (#38287)
The widgetBuilder.spec.tsx file is broken up to allow better 
parallelization of tests in CI (we currently parallelize files).

The widgetBuilder.spec.tsx test file now
contains tests that test the generic interactions
between most components in the WidgetBuilder. Tests for the
SortBy step can be found in (and should be added to)
./widgetBuilderSortBy.spec.tsx and tests for specific dataset
behaviour can be found (and should be added to) ./widgetBuilderDataset.spec.tsx.