ref(indexer): allow spanner indexer in prod via indexer_db cli arg (#37745)
**context**
We want to test out alternative backends for the metrics indexer, and test them in parallel to the current indexer. I've added a cli arg `--indexer-db` that defaults to `postgres`, that can be used to determine which backend we want to use. 

There is no problem with having another consumer group consume from the ingest topic (`ingest-metrics` or `ingest-performance-metrics`). However, we don't want to double produce to the snuba topic so we have to make a new topic to produce to. I've only added the logic for the performance metrics, but the same could be done for release health (aka would need a new dummy topic and logic to override the `output_topic`)

**other changes**
* ~I got rid of the `Db.Key` because it looked like that mapping could be replaced by using the `use_case_id`. Open to feedback on that, can always put it back.~ Got pulled out by @lynnagara  in https://github.com/getsentry/sentry/pull/37790
* I had to update the `WritesLimiter` but I think that makes sense so that we have a separation in redis between the backend implementations 
    * **[edit]** update rate limiter in https://github.com/getsentry/sentry/pull/38192
* I pass the full `MetricsIngestConfiguration` to the `process_messages` now instead of just the `use_case_id`. Seemed to make sense since both of the args I would have been passing in were on the config already
    *  **[edit]:** update the message processor in https://github.com/getsentry/sentry/pull/38225

This _should_ be able to be run in prod provided the ops config for the consumer and topic are added. 

**notes:**
The Last Seen Updater currently only works with the postgres backend as it is coupled to the postgres tables, futures changes would need to be made to enable different backends for the last seen updater https://github.com/getsentry/sentry/blob/4f0040a2f713d29c9eb17cf19a2574e1c97da107/src/sentry/sentry_metrics/consumers/last_seen_updater.py#L128