chore: Colocate tests for actionCreators, utils and components (#38339)
An outcome from Frontend TSC, we're co-locating
tests to make them more discoverable, have them as
references to learn about the module/components
and better visibility into what is being tested. [This](https://github.com/getsentry/sentry/pull/37618/files)
change should ensure tests under the test/ folder
as well as colocated tests get run in CI as we move
files over in batches.

This PR also rebalances the jest tests.

Bigger change than the previous ones just because
it's a pain to balance the jest tests, but looks like
it has to be done for every PR.


(PR 8/9*)
Previous PRs:
https://github.com/getsentry/sentry/pull/38234
https://github.com/getsentry/sentry/pull/38235
https://github.com/getsentry/sentry/pull/38240
https://github.com/getsentry/sentry/pull/38248
https://github.com/getsentry/sentry/pull/38251
https://github.com/getsentry/sentry/pull/38260
https://github.com/getsentry/sentry/pull/38306

*I think _one_ more PR should be enough to
move the rest of the files!