fix(replays): Improve console row hovering (#38095)
Adds a purple line below console messages if the `currentTime` (purple300) or `hoverTime` (purple200) is between console message timestamps.

If the currentTime or hoverTime are exact matches to the console message then you get a purple border on the left of the console message instead of the line below. "exact" in this case is based on it being the same second, not millisecond precision. You'll see this left-border most often when you hover over the console message directly.

Fixes #36014