open 'read the docs' button in a new window
While searching for a key you can open the docs in a new window. Once you find what you're looking for, you are back in the focused search bar state that you left from, instead of the current flow of going back in browser history and reloading the page you came from.