feat: Support events and transactions topics being on seperate clusters (#38278)
We are currently in the process of splitting the events and transactions topics.

The previous version of the post process forwarder assumed that `events`
and `transactions` were either using the same Kafka topic (as was the case
previously) or were different topics in the same cluster. It wouldn't work
properly if the two topics were in different Kafka clusters. This change drops
that requirement, and it enables us to move transactions to a new cluster to reduce
load on the main events cluster.