feat(perf-issues): Issue creation project option UI (#38472)
### Summary
This adds a range slider we can use to opt-in a project to perf issues, which will help during rollout. UI is behind the performance-issues flag. Using a float for now for rollout, this can be switched to 'enabled by default'/boolean later if needed. Will follow up and add thresholds etc. but getting this blocking part in first.

### Screenshots

![Screen Shot 2022-09-06 at 1 58 02 PM](https://user-images.githubusercontent.com/6111995/188706255-c26000ca-16f1-477f-b1bd-6b4f66c02e93.png)