feat(eventstream): Support transitioning a subset of project IDs to a new topic
This introduces a new killswitch `kafka.send-project-transactions-to-new-topic`.
Project IDs which are added here will be sent to the new topic (KAFKA_NEW_TRANSACTIONS
instead of KAFKA_TRANSACTIONS).

This switch is intended to be temporary code to support a gradual SaaS rollout of
the events/transactions topic split. Once all projects are moved to the new pipeline
we should remove this code.