feat(perf_issues): Add `GroupEvent` and split some functionality in `Event` into a base class. (#38143)
Since we can now have events with multiple groups, we can no longer rely on the `Event.group`
property. This pr adds in a `GroupEvent` subclass that should be passed around wherever we expect an
event to have a single `Group` associated with it.

`Event` has been split up into `BaseEvent` and `Event`. We will deprecate and remove uses of
`group_id` and `group` in the `Event` class going forward. If we need an event with a `Group`, we
can use `build_group_events` to fetch all `GroupEvents` associated with the `Event`, or `for_group`
if we just need a specific `Event`/`Group` pairing.

Going forward, the plan is to store all groups in the `groups` property. This means that error
events being sent via eventstream will have their group included in `groups` as well. We'll
need to update the errors processor in snuba to look there instead of `group_id`. This seems cleaner
long term, instead of having both `group_id` and `group_ids` passed through.

To figure out where we need to use `build_group_events` and `for_group` we can do a mix of searching
the codebase and commenting  out the `group_id` and `group` properties and see how CI goes.