feat(perf-issues): Add N+1 DB Span detector (#38372)
We had a generic identical spans detector which should also detect N+1s,
but this is a much more targeted approach especially focused on getting
good fingerprints.

We are strict about the structure we look for: we must be able to
identify a "source" span (the DB call precipitating the N+1) in addition
to the N identical spans, and all those spans must be under a common
parent. All of that information is used in the fingerprint so that we
can group issues correctly.

For now, this detector will be the only one outputting problems for
downstream performance issue creation.