feat(perf-issues): Add settings endpoint for detector options (#38527)
The detector options will be far more numerous that the basic all issue
rate previously added to the project detail endpoint. Made a new
endpoint to get/set into the project option that will contain all the
settings as a flat dict, making use of a serializer to do validation.