feat(metrics): Support field aliasing on metrics layer [TET-322 TET-356] (#38509)
Adds logic that allows for aliasing on all instances of `MetricField`.
If `MetricField` references an metrics expression or an instance of
`SingleEntityDerivedMetric` then, the alias is propagated down to query
level, and the response has the aliased names. If however, it references
an instance of `CompositeEntityDerivedMetric`, then it creates alias
that corresponds to the mri of the constituent suffixed with
`__CHILD_OF__<parent_alias>`, and then post query the value of the
composite derived metric is computed from the children values. This
alias suffixing behaviour allows us to avoid naming collisions when
requesting the same composite dervied metric more than once but with
different arguments. This PR also adds meta type for composite entity
derived metric.

ToDo:
- Will handle aliasing groupBy aliasing in a subsequent PR