feat(perf-issues): Add non-flagr feature flag for ingest
This splits out the creation/ingest portion of performance issues into a non-flagr feature flag, since we don't want to use flagr in the ingest pipeline for speed / stability purposes. This feature exists for the final internal / EA / GA rollout, and when an organization is opted in via a separate getsentry PR they will start having performance issues created immediately (since we are defaulting performance issue creation to 'On' for rollout, given it shouldn't have a cost impact).