feat(issue_search): Allow issue search to run only on Postgres (#38570)
We have code in place to allow us to make queries just to Postgres when
doing issue search. This code path is only allowed when we don't pass in
an `end` date, which means it is never used on the issue list.

It should be fine to just rely on Postgres here. We're already making
these queries to Postgres to just fetch the IDs, so it shouldn't cause
extra load. To be safe, I'm putting this behind a feature flag so that
we can roll it out safely on Sentry first.

I've also switched the query to use a replica, which we do for fetching
group id candidates later on. This should help us avoid putting extra
load on our primary Postgres.

There are two main benefits here to this:
 - Reduce load we send through to Snuba unnecessarily
- Allow performance issues to appear on the issue list before we can
query them correctly in Snuba