fix(metrics): Fix startup crash in parallel indexer [sns-1490] (#38455)
Since https://github.com/getsentry/sentry/pull/38225 the parallel
indexer fails to serialize the processing function here:
https://github.com/getsentry/sentry/blob/9bf499ad95030ed1112f117c5c1be59b2e036509/src/sentry/sentry_metrics/consumers/indexer/parallel.py#L115

We need to make sure the message processor is pickleable. So the config
also needs to be pickleable.

The old code worked because it imported the config and indexer from
settings instead of attempting to pickle them.