feat(post-process-forwarder): Use correct cluster if topic name override passed
This change ensures that the right Kafka cluster is used if a topic name override
is passed via CLI. Previously this code did not work properly if the topic was not
on the same cluster as the default.

This change is needed for the roll out of the errors/transactions split. We will
temporarily run two transactions post process forwarders side by side
(one consuming from the old topic and one from the new one). The new topic will
be in a different cluster.