feat(perf_issues): Store problem hashes on transactions, and store problem details in nodestore
This pr makes sure we store all relevant problem hashes on the transaction, and stores the detailed
evidence of the problem in a separate key in nodestore.

We only store problem hashes that are actually used, so anything that was exluded due to rate
limiting will not be stored against the event. Existing hashes are always stored.

We also store a serialized version of `PerformanceProblem` in nodestore for all the used hashes.
Again, things that were ratelimited won't be written here.

We don't need to explicitly ttl values that we set in nodestore. Ttl is handled behind the scenes
for us. Notes on investigation:
Nodestore when backed by bigtable is set up in prod to always have a ttl of 90 days - we don’t need
to set that on the key explicitly.
https://github.com/getsentry/getsentry/blob/142cb40e3eee51fe08f953963cd869ef8a5834d4/getsentry/conf/settings/prod.py#L301-L308

When backed by Postgres (IE self hosted), the ttl is handled by the cleanup script. Note that this
script does nothing when the store is bigtable, since we don’t have cleanup implemented.
https://github.com/getsentry/sentry/blob/a7fec2d12145b2840499e757cf3d50d725addd63/src/sentry/runner/commands/cleanup.py#L265-L268
https://github.com/getsentry/sentry/blob/a7fec2d12145b2840499e757cf3d50d725addd63/src/sentry/nodestore/django/backend.py#L54-L62