fix(discover): Fixes mangled drilldown links in Discover aggregated results tables (#38468)
Better detects erroneous or/and operators due to parenthesis and also
removes negated filters on the supplied key.
Previously, when drilling down on a field `b` with query filters `a:a1
(b:b1 OR b:b2) c:c1`, the result of `removeFilter` would be `a:a1 OR
c:c1` which is incorrect. This PR fixes it by correcting the result to
`a:a1 c:c1`.