feat(dashboards): Allow saving Custom Performance Metrics widgets using Discover dataset fallback (#38523)
Previously we would alert the user and disable saving widgets if they
attempt to create a Custom Performance Metric widget with settings
incompatible with the Metrics dataset. This update removes that
restriction and allows Custom Perf Metric widgets to be created using
the Discover dataset as a fallback if settings are incompatible with
metrics data.