feat(hybrid-cloud): Introduce Region info and sentry region config (#38655)
We will be storing region information for hybrid cloud in a
configuration. In this repo, the configuration will be blank allowing
everything to work out of the box. Region information will be saved as a
new Region dataclass.

Note that any new fields added to the Region dataclass must be kwargs to
maintain backwards compatibility.