feat(perf-issues): track N+1s dropped due to query truncation
We know that the Python SDK (and possibly others) truncates the span
description field, so we don't always have full SQL queries available to us.
Previously, we failed early when encountering a truncated query. Instead, run
the detector right up to when it would report an N+1, and only then check for
query truncation. That way we can record an accurate metric of how many times
we _would_ have reported an N+1 but _couldn't_ because of truncation.