feat(backend): adds default_owner_id (#38520)
This PR creates a new property called `default_owner_id` on the
`Organization` which is similar to `get_default_owner`. The main
difference is that if for some reason there is no owner for an
organization, we return `None` instead of throwing a `IndexError`.

As for migration of `get_default_owner`, that's tricky because some
places are catching an `IndexError` so we can't just do a drop-in
replacement.