fix(profiling): do not forward unwanted qs (#38740)
Some parts of the query string collide between the flamechart views
(query) and some break the API (cursor). Make sure we strip these out
when navigating between the pages. This is more brittle than I would
have wanted it to be, but trying to preserve as many qs as we can makes
for a nicer back/forth navigation as state is preserved.

We should rethink how we handle these if it proves to be too brittle.