feat(perf-issues): Add system-wide options for thresholds (#38742)
This adds some system options for the detection thresholds that will be
used as the defaults instead of project-option defaults, unless the
project-options are set.

They will be used for internal/EA etc. out to GA for rollout, and will
be paired with the 'performance-issues-ingest' non-flagr feature flag to
automatically opt organizations onto perf-issues.

These options need to be put in and initialized before we bind them to
the existing flow.