refs: Remove unused group serializer and unused endpoints that use `StreamGroupSerializer` (#38725)
Maintaining `StreamGroupSerializer` is painful. It's not really used in
many places and there's not a clear reason for why it's different.

The endpoints it is used in are mostly completely unused, I confirmed
this in datadog. Removing these endpoints and
`TagBasedStreamGroupSerializer`. Hopefully we can do a little more work
and kill `StreamGroupSerializer`.