fix(metrics): Ability to configure releasehealth cluster
We are not going to configure any releasehealth quotas, but still need
to be able to connect to a valid redis cluster if we wanted to do so.

We probably don't want to get rid of all the limiting code related to
releasehealth because that's the only way we can test things in the QE
environment.