fix(api): Fix N+1 organization queries in OrganizationTeams endpoint
Detected as a Performance Issue. The organization is loaded once per team
returned by the OrganizationTeams endpoint, and in the detected case, the calls
totalled 368ms. (That is unusually slow for a simple lookup by ID, but even if
the DB is uncharacteristically slow, we should still see net improvement by
grouping into a single call.

Fixes SENTRY-VZG