feat(hybrid-cloud): Add an organization region map table
This table maps an organization to a region key. This region key is then used with SENTRY_REGION_CONFIGS to retrieve region details.

This table is a workaround for an issue that prevents new columns in the Organization table: https://github.com/getsentry/sentry/issues/38821

Once that issue is fixed a new migration should merge this table and the Organizaiton table