fix(perf-issues): N+1 with odd number of potential sources missed
There was a bug in the N+1 DB Queries detector that would sometimes miss valid
N+1s.

When reacting to a sequence of `db` spans, it was effectively only treating
every _second_ db span as a potential N+1 source.

When a span breaks an N+1 pattern, not only should the current span be
considered as a potential source, but so should the last span in `n_spans` -
since that span never got a chance at being a source span the first time
through.