feat(perf_issues): Default `sentry:performance_issue_creation_rate` to 1.0, and gate perf issue creation using `performance-issues-ingest` (#38864)
We want to enable performance issues for EA. This means that we should
default the project level setting to 1.0. Then we just gate creation
behind `organizations:performance-issues-ingest`, which makes sure this
will be enabled only for EA customers as we roll it out to them.

I think that at this point we could also remove the
`performance.issues.all.problem-creation` option and just rely on this
feature flag. It might also make sense to only check the
`organizations:performance-issues-ingest` flag as part of ingest and
move it out of the detection code.