ref(sentry-perf): Add span to missing instrumentation (#38963)
Adds a span to the missing instrumentation for `/issues` endpoint.

This is an example of using profiling to find areas to improve spans,
the actual fix likely has to do with ThreadExecutor and our sdk
integration since there are snuba queries being run, which likely have
unpropagated traces. We can fix that later.