fix(migrations) Fix migration generation for organization
Generating migrations for organization was failing because the feature
check would hydrate a record causing a query which attempted to use the
new schema state. By catching that error and defaulting the feature to
off we can successfully generate migrations for organization again.

I didn't include tests as I couldn't think of a good way to simulate
adding a column that isn't in the schema yet in the test suite.

Fixes #38821