ref(MDC): Update Snuba eventstream and insert endpoints (#38877)
This PR is responsible for updating the Snuba eventstream and insert
test endpoints
* From: `POST /tests/{dataset}/eventstream` To: `POST
/tests/{entity}/eventstream`
* From: `POST /tests/{dataset}/insert` To: `POST
/tests/entities/{entity}/insert`

This change is required due to the removal of the `default_entity`
attribute in the Dataset class in Snuba. As a result, every request sent
to Snuba (e.g. query, eventstream, etc.) must specify both a dataset and
entity.
The changes to the server endpoint is specified here:
https://github.com/getsentry/snuba/pull/3142