ref(perf): Start replacing *DiscoverQuery with RQ (#38025)
This is an example of replacing our fetch-as-component pattern with
hooks, as was eventually intended. In this case using React Query to try
it out on the performance part of the product, since almost every single
request we make is through *DiscoverQuery so it would be easy to test it
out with this hook and swap up React Query for another approach under
the hood if we'd like to at a later point.

Since this is touching our bundle sizes, you know the drill:
- Around ~**10kb** for pages that choose to use this chunk. Given our low
correlation of asset size to lcp value this should in theory have no
impact on our LCP etc. for these pages.