ref(issues): Improve grouping types (#39098)
Some improvements to `Group*` TypeScript types in preparation for
performance events grouping UI.

- split up variant types, and make a union type out of their
combination. This is more specific, and allows for type narrowing. With
type narrowing there's no need to sprinkle `?.` everywhere, plus this
now matches the variants defined in Python, which is nice
- add missing variants (checksum, fallback) which sometimes do come back
from the backend
- change `?.` lookups to `in` checks. `in` narrows types in TypeScript,
`?.` does not narrow types