feat(debug): Add send button to email previews (#39093)
![image](https://user-images.githubusercontent.com/8980455/191379141-d869824f-3954-481c-aa13-265a8f32cb66.png)

Added a button to email debug views. Once clicked it'll send the email
using the default backend.

Note that as of right now email debug views are somewhat messy. some
debug views are functional while others are classes. This feature
requires inheriting from `MailPreviewView` on a per-view basis. This
will be addressed in a future PR.