feat(perf-issues): Add extended n+1 db detector
This extends our existing n+1 db detector so we can try to improve detection without creating more issues (for the moment). For the time being this is detection-only, and for the first change we'll get rid of the 'parent' requirement not allowing for the root span. A lot of our DB integrations don't have parent information, or don't have extra spans to put under the root of the transaction that would act as the parent. This new detector should work for 'nodejs (mongo / pg / more)' and ruby without needing to fix the db.span situation.