feat(commit-context): Create analytic event groupowner.assignment (#39118)
## Objective:
We want to track how many times the Commit Context groupowner assignment
got changed for the same group. This should indicate that the stacktrace
had changed between the events and should be proof that our assumption
that the grouping algorithm is accurate is false. The caveat for this
metric is that it will only be created if a new event occurs 7 days
after the original assignment