feat(metrics): Add support for counting transactions with condition on value [TET-326] (#38938)
This PR extends the Metric MRI layer with a new DerivedOp called count_transaction_with_condition which generalizes over the operations found here.

The operation implemented enables the calculation of the number of occurrences of a given metric with a transaction value equal to << unparameterized >> or "" or neither of the two. The transaction value refers to the value of tag transaction.