feat(perf-issues): Add extended n+1 db detector (#39161)
This extends our existing n+1 db detector so we can try to improve
detection without creating more issues (for the moment). For the time
being this is detection-only, and for the first change we'll get rid of
the 'parent' requirement not allowing for the root span. A lot of our DB
integrations don't have parent information, or don't have extra spans to
put under the root of the transaction that would act as the parent.

This new detector should work for `nodejs (mongo / pg / more)` and
`ruby` without needing to fix the db-spans-without parents issue.

#### Rollout
- This is currently detection only, but I've diverged from the original
behaviour of a 1:1 detector to group type relation. This new detector
still related to the n+1 db group, since eventually it will replace it.
- We should (in an upcoming PR) add a duplicate check for performance
problems before passing off to be grouped.
- We can have a separate system option for the creation rate on this
detector and turn down the original and turn up this extended detector.
Since the fingerprint should be the same, as well as the perf-type for
existing issues, there shouldn't be more created groups. Alternatively
we can just put a PR to switch the class for detectors.