Elias/fix/mock api client warn mode (#39289)
This is a follow up PR to this:
https://github.com/getsentry/sentry/pull/38769

tldr; mock client errors were being gobbled up by a `setTimeout` and in
certain scenarios where we had alot of tests we would associate mock
client errors with the wrong test. this PR enforces that we error if we
are missing a mock response. for existing tests which are missing mocks
we've exposed a `warnOnMissingMocks` mode. this method is marked as
`@deprecated` and should show up with a strikethrough/message w/
intellisense

![image](https://user-images.githubusercontent.com/7349258/192304820-eb649a4f-9e1d-484f-aa4f-d932f60d08e3.png)