feat: Add arroyo compatible synchronized consumer
This brings over the synchronized consumer code that was previously
part of arroyo. It was removed in https://github.com/getsentry/arroyo/pull/81
since it was no longer being used by Snuba and should not live in the
shared library.

This implementation differs from the current synchronized consumer code
however since it is functionally the same. This code has been fairly extensively
tested (as it was previously used with Snuba subscriptions)
It is currently not used but will be used when converting the post process
forwarder to use Arroyo.