fix(discover): Flicker due to fallback to stale savedQuery
When the user clicks "Save Changes", we redirect them to a short URL. At
this point, savedQuery from the AsyncComponent is stale. The Results
component then tries to use this stale query to build an eventView, but
since we've redirected to a short URL the changes aren't there to
override the values properly.

Instead, when the update happens successfully, save the response so we
derive the proper eventView.